requirements.txt
# Core dependencies
langchain
langchain-community
langchain-huggingface
langchain-pinecone
langchain-openai
langchain-core
langgraph

# Optional: For reranking (uncomment if needed)
flashrank

# Hugging Face embeddings
sentence-transformers
torch

# Vector database
pinecone-client

# Environment management
python-dotenv

# Pydantic for structured output
pydantic

# Utilities
tqdm

